% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gauss_1D_sample}
\alias{gauss_1D_sample}
\title{Univariate sample}
\format{
A vector with 100 rows
}
\usage{
data("gauss_1D_sample")
}
\description{
A univariate sample of 100 realisations of a gaussian law with mean 0 and standard deviation 1
}
\keyword{datasets}
