% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gauss_mD_sample}
\alias{gauss_mD_sample}
\title{Multivariate sample}
\format{
A matrix with 100 rows and 2 columns
}
\usage{
data("gauss_mD_sample")
}
\description{
A 2D sample of 100 realisations of a gaussian law with mean (0, 0) and covariance matrix
| 1    0.9 |
| 0.9  1   |
}
\keyword{datasets}
